requirements.txt
praw>=7.7
requests>=2.28
python-dotenv>=1.0
discord.py>=2.3
fastapi>=0.110
uvicorn>=0.23
